fix putting args into memory

+++ 03-code/headers/stddef.h
@@ -4,7 +4,7 @@
 #define NULL ((void*)0)
 
 typedef unsigned long size_t;
-typedef long ptrdiff_t;
+typedef int ptrdiff_t;
 typedef int wchar_t;
 
 #endif
